+++ frontend/src/components/core/Block/styled-components.ts
@@ -106,13 +106,3 @@ export const StyledBlock = styled.div<StyledBlockProps>(
     }
   }
 )
-
-export interface StyledFormProps {
-  width: number
-}
-
-export const StyledForm = styled.div<StyledFormProps>(({ width }) => ({
-  outline: "1px solid",
-  padding: "1em",
-  width,
-}))
